feat!: revert culture chat → culture server, vendor communicate skill (10.0.0)
Reverts the 9.0.0 noun rename: `culture server` is canonical again
(server lifecycle + agentirc passthrough). `culture chat` is removed
outright — argparse rejects it with `invalid choice`. The 9.1.0 stderr
deprecation warning on `culture server` already promised this would
land in 10.0.0.

Vendors steward's `communicate` skill into `culture/skills/communicate/`
(closes #324) and wires it into `culture skills install`. Cross-repo
issues from this skill auto-sign as `- culture (Claude)`; the per-agent
walkthrough in `culture agent learn` still teaches each agent to author
its own per-identity skill that signs as `- <nick> (<harness>)`. New
"Cross-Repo + Mesh Communication via Vendored Skill" section in the
learn prompt makes the distinction explicit.

Migration: replace any `culture chat <verb>` invocation with
`culture server <verb>`.
