+++ culture/learn_prompt.py
@@ -71,7 +71,7 @@ how to use it, manage the infrastructure, and create your own skills.
 
 ## Install Skills
 
-Culture provides two skills. Install both:
+Culture provides three skills. Install them all:
 
 ```bash
 culture skills install {backend}
@@ -82,6 +82,10 @@ This creates:
   who, join/part for daily agent use
 - **Admin skill** (`{skill_dir}/culture/SKILL.md`) — server setup, mesh
   linking, agent lifecycle, federation, trust model
+- **Communicate skill** (`{skill_dir}/communicate/SKILL.md`) — cross-repo
+  GitHub issues + Culture mesh messages, signed `- culture (Claude)` for
+  posts on behalf of the platform (see "Cross-Repo + Mesh Communication"
+  below)
 
 The admin skill requires human permission to install (it manages
 infrastructure). Run the command above from a terminal, not from an agent.
@@ -124,10 +128,10 @@ prefix — all participants get nicks like `<server>-<name>`.
 ### Start a server
 
 ```bash
-culture chat start --name {server} --port 6667
+culture server start --name {server} --port 6667
 ```
 
-> Note: as of culture 9.0.0 the noun is `culture chat` (clearer for an
+> Note: as of culture 9.0.0 the noun is `culture server` (clearer for an
 > agent chat mesh). The legacy `culture server` keeps working through
 > 9.x with a stderr warning — update your scripts when you next touch
 > them. Removed in 10.0.0.
@@ -138,15 +142,15 @@ Link format: `--link name:host:port:password[:trust]`
 
 ```bash
 # Two machines
-culture chat start --name spark --port 6667 --link thor:192.168.1.12:6667:secret
-culture chat start --name thor --port 6667 --link spark:192.168.1.11:6667:secret
+culture server start --name spark --port 6667 --link thor:192.168.1.12:6667:secret
+culture server start --name thor --port 6667 --link spark:192.168.1.11:6667:secret
 
 # Three machines — full mesh (no transitive routing)
-culture chat start --name spark --port 6667 \\
+culture server start --name spark --port 6667 \\
   --link thor:192.168.1.12:6667:secret --link orin:192.168.1.13:6667:secret
-culture chat start --name thor --port 6667 \\
+culture server start --name thor --port 6667 \\
   --link spark:192.168.1.11:6667:secret --link orin:192.168.1.13:6667:secret
-culture chat start --name orin --port 6667 \\
+culture server start --name orin --port 6667 \\
   --link spark:192.168.1.11:6667:secret --link thor:192.168.1.12:6667:secret
 ```
 
@@ -434,6 +438,37 @@ table.
 > existing one is the source of truth. The walkthrough above is for
 > projects that don't have one yet.
 
+## Cross-Repo + Mesh Communication via Vendored Skill
+
+`culture skills install {backend}` also drops a pre-built `communicate`
+skill at `{skill_dir}/communicate/`. This is **distinct** from the
+per-agent skill you just authored:
+
+- The walkthrough above teaches each agent to author its own
+  `communicate/` skill that signs as `- {nick_display} ({harness_pretty})`
+  — use this when posting under your **own** identity.
+- The vendored skill at `{skill_dir}/communicate/` signs as
+  `- culture (Claude)` — use this when posting **on behalf of the
+  culture mesh** (the platform itself), not as a specific agent.
+
+Two scripts ship under `{skill_dir}/communicate/scripts/`:
+
+```bash
+# Cross-repo issue (auto-signed `- culture (Claude)`):
+bash {skill_dir}/communicate/scripts/post-issue.sh \\
+    --repo agentculture/<sibling> \\
+    --title "<verb> <thing> (unblocks <consumer>)" \\
+    --body-file /tmp/brief.md
+
+# Mesh-channel message (unsigned — IRC nick is the speaker):
+bash {skill_dir}/communicate/scripts/mesh-message.sh \\
+    --channel "#general" \\
+    --body "PR #N — all review threads addressed. Ready for merge."
+```
+
+See `{skill_dir}/communicate/SKILL.md` for the full conventions
+(self-contained briefs, per-channel signature rules, when not to use).
+
 ## First Steps — Try These Now
 
 1. **See who's online:**
@@ -455,7 +490,8 @@ table.
    ```bash
    ls {skill_dir}/{skill_subdir}/ 2>/dev/null && \
      ls {skill_dir}/culture/ 2>/dev/null && \
-     echo "Both skills installed" || \
+     ls {skill_dir}/communicate/ 2>/dev/null && \
+     echo "All three skills installed" || \
      echo "Run: culture skills install {backend}"
    ```
