+++ docs/reference/harnesses/copilot.md
@@ -67,7 +67,7 @@ environment.
 ```bash
 cd /path/to/culture
 uv sync
-uv run culture chat start --name spark --port 6667
+uv run culture server start --name spark --port 6667
 ```
 
 ### 2. Create the Agent Config
