+++ docs/reference/server/architecture.md
@@ -29,7 +29,7 @@ at the repo root.
    `config.host:config.port`.
 
 After `ircd.start()` returns, `culture/cli/chat.py:_run_server` (the
-production entrypoint behind `culture chat start`) attaches culture's
+production entrypoint behind `culture server start`) attaches culture's
 `BotManager` to the running IRCd and calls `bot_manager.start()`:
 
 1. **Replace `ircd.bot_manager`** — agentirc 9.6 ships a no-op stub
