+++ tests/test_server_shim.py
@@ -1,4 +1,4 @@
-"""Pin the `culture chat` command tree shape after Phase A3.
+"""Pin the `culture server` command tree shape.
 
 Two halves:
 
@@ -8,9 +8,12 @@ Two halves:
 
 2. **Forwarded verbs** (`restart`, `link`, `logs`, `version`, `serve`)
    pass through to `agentirc.cli.dispatch` verbatim. The simplest stable
-   assertion: invoking `culture chat <verb>` and `agentirc <verb>`
+   assertion: invoking `culture server <verb>` and `agentirc <verb>`
    produces the same exit code for the same args (and the version
    command prints the same string).
+
+`culture chat` was removed in 10.0.0 — it should not be a recognized
+subcommand anymore.
 """
 
 from __future__ import annotations
@@ -20,7 +23,7 @@ import sys
 
 import pytest
 
-CULTURE = [sys.executable, "-m", "culture", "chat"]
+CULTURE = [sys.executable, "-m", "culture", "server"]
 # Invoke via `python -m agentirc` rather than the `agentirc` console
 # script so the test works in environments where entry-point scripts
 # aren't on PATH (CI containers, Windows, isolated venvs).
@@ -43,11 +46,11 @@ def _run(cmd: list[str]) -> tuple[int, str, str]:
     return proc.returncode, proc.stdout, proc.stderr
 
 
-def test_chat_top_level_help_lists_every_verb() -> None:
+def test_server_top_level_help_lists_every_verb() -> None:
     rc, out, _ = _run([*CULTURE, "--help"])
     assert rc == 0
     for verb in (*CULTURE_OWNED_VERBS, *FORWARDED_VERBS):
-        assert verb in out, f"verb {verb!r} missing from `culture chat --help`"
+        assert verb in out, f"verb {verb!r} missing from `culture server --help`"
 
 
 @pytest.mark.parametrize("verb", CULTURE_OWNED_VERBS)
@@ -62,12 +65,12 @@ def test_culture_owned_verb_help_dispatches_locally(verb: str) -> None:
 
 
 def test_forwarded_version_matches_agentirc() -> None:
-    """`culture chat version` and `agentirc version` print the same line."""
+    """`culture server version` and `agentirc version` print the same line."""
     cul_rc, cul_out, _ = _run([*CULTURE, "version"])
     agent_rc, agent_out, _ = _run([*AGENTIRC, "version"])
     assert cul_rc == agent_rc == 0
     assert cul_out.strip() == agent_out.strip(), (
-        "culture chat version drifted from agentirc version: " f"{cul_out!r} vs {agent_out!r}"
+        "culture server version drifted from agentirc version: " f"{cul_out!r} vs {agent_out!r}"
     )
 
 
@@ -76,9 +79,22 @@ def test_forwarded_verb_is_reachable(verb: str) -> None:
     """Each forwarded verb should be reachable; --help for those that
     accept it should exit 0. agentirc-owned argparse may not accept
     `--help` for every verb, so we only assert the verb is reachable
-    (no `Unknown chat command` error)."""
+    (no `Unknown server command` error)."""
     _, out, err = _run([*CULTURE, verb, "--help"])
     combined = (out + "\n" + err).lower()
     assert (
-        "unknown chat command" not in combined
+        "unknown server command" not in combined
     ), f"forwarded verb {verb!r} not reaching agentirc.cli.dispatch"
+
+
+def test_culture_chat_is_removed() -> None:
+    """`culture chat` was removed in 10.0.0 — argparse should reject it."""
+    proc = subprocess.run(
+        [sys.executable, "-m", "culture", "chat", "--help"],
+        capture_output=True,
+        text=True,
+        timeout=30,
+    )
+    # argparse exits 2 on unrecognized subcommand.
+    assert proc.returncode != 0
+    assert "invalid choice" in proc.stderr.lower() or "chat" in proc.stderr.lower()
